Add a geli resize subcommand to resize encrypted filesystems prior to growing the filesystem.
Refuse to attach providers where the metadata provider size is
wrong.  This makes post-boot attaches behave consistently with
pre-boot attaches.  Also refuse to restore metadata to a provider
of the wrong size without the new -f switch.  The new -f switch
forces the metadata restoration despite the provider size, and
updates the provider size in the restored metadata to the correct
value.

Helped by:	pjd
Reviewed by:	pjd


git-svn-id: http://svn.freebsd.org/base/head@212934 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f