sh: Fix some issues with aliases and case, by importing dash checkkwd code.
This moves the function of the noaliases variable into the checkkwd
variable. This way it is properly reset on errors and aliases can be used
normally in the commands for each case (the case labels recognize the
keyword esac but no aliases).

The new code is clearer as well.

Obtained from:	dash


git-svn-id: http://svn.freebsd.org/base/head@214709 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f